warning reduction

+++ combat/CombatDamage.cpp
@@ -106,7 +106,7 @@ namespace {
         if (target_ships)
             retval.push_back(fighter_damage * fighter_shots);
         else
-            retval.push_back(fighter_shots);
+            retval.push_back(static_cast<float>(fighter_shots));
         return retval;
     }
 
@@ -251,8 +251,8 @@ int Combat::TotalFighterShots(const ScriptingContext& context, const Ship& ship,
 {
     // Iterate over context, but change bout number
     ScriptingContext mut_context{context};
-    int launch_capacity = ship.SumCurrentPartMeterValuesForPartClass(MeterType::METER_CAPACITY, ShipPartClass::PC_FIGHTER_BAY, context.ContextUniverse());
-    int hangar_fighters = ship.SumCurrentPartMeterValuesForPartClass(MeterType::METER_CAPACITY, ShipPartClass::PC_FIGHTER_HANGAR, context.ContextUniverse());
+    int launch_capacity = static_cast<int>(ship.SumCurrentPartMeterValuesForPartClass(MeterType::METER_CAPACITY, ShipPartClass::PC_FIGHTER_BAY, context.ContextUniverse()));
+    int hangar_fighters = static_cast<int>(ship.SumCurrentPartMeterValuesForPartClass(MeterType::METER_CAPACITY, ShipPartClass::PC_FIGHTER_HANGAR, context.ContextUniverse()));
     int launched_fighters = 0;
     int shots_total = 0;
     Condition::ObjectSet condition_matches;
